refactor: AddPicture 페이지 props 기반으로 리팩토링
