+++ src/components/pages/guestbook/AddPicture/AddPicture.tsx
@@ -1,22 +1,18 @@
 import styles from "@/components/pages/guestbook/AddPicture/AddPicture.module.css";
 import Add from "@/components/pages/guestbook/AddPicture/Add";
-import { useState } from "react";
 import { IoMdClose } from "react-icons/io";
 
-export default function AddPicture() {
-  const [images, setImages] = useState<string[]>([]);
+interface AddPictureProps {
+  images: string[];
+  onAdd: (url: string) => void;
+  onRemove: (index: number) => void;
+}
 
-  const handleAdd = (url: string) => {
-    if (images.length >= 4) {
-      return alert("최대 4개까지만 업로드 가능합니다!");
-    }
-    setImages((prev) => [...prev, url]);
-  };
-
-  const handleRemove = (index: number) => {
-    setImages((prev) => prev.filter((_, i) => i !== index));
-  };
-
+export default function AddPicture({
+  images,
+  onAdd,
+  onRemove,
+}: AddPictureProps) {
   return (
     <div className={styles.container}>
       <div className={styles.box}>
@@ -28,24 +24,27 @@ export default function AddPicture() {
         <div className={styles.picture}>
           {[0, 1, 2, 3].map((slotIndex) => {
             const image = images[slotIndex];
+
             if (image) {
               return (
                 <div key={slotIndex} className={styles.imageBox}>
                   <img src={image} alt={`upload-${slotIndex}`} />
                   <IoMdClose
                     className={styles.cancelIcon}
-                    onClick={() => handleRemove(slotIndex)}
+                    onClick={() => onRemove(slotIndex)}
                   />
                 </div>
               );
             }
+
             if (slotIndex === images.length) {
               return (
                 <div key={slotIndex} className={styles.imageBox}>
-                  <Add onAdd={handleAdd} />
+                  <Add onAdd={onAdd} />
                 </div>
               );
             }
+
             return <div key={slotIndex} className={styles.imageBox} />;
           })}
         </div>
